Repair tests.

+++ src/main/java/euler/p006/SumSquareDifference.java
@@ -4,16 +4,16 @@ import java.util.stream.IntStream;
 
 public class SumSquareDifference {
 
-    public static int calculateDifference(int limit) {
+    public static long calculateDifference(long limit) {
         return calculateSquareOfTheSum(limit) - calculateSumOfTheSquares(limit);
     }
 
-    public static int calculateSumOfTheSquares(int limit) {
-        return IntStream.rangeClosed(1, limit).map(n -> n * n).sum();
+    public static long calculateSumOfTheSquares(long limit) {
+        return IntStream.rangeClosed(1, (int) limit).map(n -> n * n).sum();
     }
 
-    public static int calculateSquareOfTheSum(int limit) {
-        int sum = limit * (limit + 1) / 2;
+    public static long calculateSquareOfTheSum(long limit) {
+        long sum = limit * (limit + 1) / 2;
         return sum * sum;
     }
 }
